requirements.txt
streamlit
tqdm
spacy
chromadb
sentence-transformers
pandas
numpy
python-dotenv
PyMuPDF
groq
